move workers
